backend/requirements.txt
fastapi
uvicorn[standard]
gunicorn
groq
python-dotenv
pydantic
xhtml2pdf
jinja2
python-multipart
httpx
motor
pymongo
pyjwt
bcrypt
pdfplumber
python-docx
slowapi
pytest
